Feat: Remove need to import `CacheConfig` classes in addition to `Cache` classes (major refactor) (#59)
