+++ airbyte/__init__.py
@@ -5,9 +5,10 @@
 """
 from __future__ import annotations
 
-from airbyte._factories.cache_factories import get_default_cache, new_local_cache
+from airbyte import caches, datasets, registry, secrets
 from airbyte._factories.connector_factories import get_source
-from airbyte.caches import DuckDBCache, DuckDBCacheConfig
+from airbyte.caches.duckdb import DuckDBCache
+from airbyte.caches.factories import get_default_cache, new_local_cache
 from airbyte.datasets import CachedDataset
 from airbyte.registry import get_available_connectors
 from airbyte.results import ReadResult
@@ -16,14 +17,20 @@ from airbyte.source import Source
 
 
 __all__ = [
-    "CachedDataset",
-    "DuckDBCache",
-    "DuckDBCacheConfig",
+    # Modules
+    "caches",
+    "datasets",
+    "registry",
+    "secrets",
+    # Factories
     "get_available_connectors",
-    "get_source",
     "get_default_cache",
     "get_secret",
+    "get_source",
     "new_local_cache",
+    # Classes
+    "CachedDataset",
+    "DuckDBCache",
     "ReadResult",
     "SecretSource",
     "Source",
