+++ airbyte/caches/_catalog_manager.py
@@ -26,8 +26,8 @@ from airbyte import exceptions as exc
 if TYPE_CHECKING:
     from sqlalchemy.engine import Engine
 
-STREAMS_TABLE_NAME = "_airbytelib_streams"
-STATE_TABLE_NAME = "_airbytelib_state"
+STREAMS_TABLE_NAME = "_airbyte_streams"
+STATE_TABLE_NAME = "_airbyte_state"
 
 GLOBAL_STATE_STREAM_NAMES = ["_GLOBAL", "_LEGACY"]
 
@@ -73,6 +73,11 @@ class CatalogManager:
         assert self._source_catalog is not None
 
     @property
+    def stream_names(self) -> list[str]:
+        """Return the names of all streams in the cache."""
+        return [stream.stream.name for stream in self.source_catalog.streams]
+
+    @property
     def source_catalog(self) -> ConfiguredAirbyteCatalog:
         """Return the source catalog.
 
@@ -116,19 +121,17 @@ class CatalogManager:
     def get_state(
         self,
         source_name: str,
-        streams: list[str],
+        streams: list[str] | None = None,
     ) -> list[dict] | None:
         self._ensure_internal_tables()
         engine = self._engine
         with Session(engine) as session:
-            states = (
-                session.query(StreamState)
-                .filter(
-                    StreamState.source_name == source_name,
-                    StreamState.stream_name.in_([*streams, *GLOBAL_STATE_STREAM_NAMES]),
+            query = session.query(StreamState).filter(StreamState.source_name == source_name)
+            if streams:
+                query = query.filter(
+                    StreamState.stream_name.in_([*streams, *GLOBAL_STATE_STREAM_NAMES])
                 )
-                .all()
-            )
+            states = query.all()
             if not states:
                 return None
             # Only return the states if the table name matches what the current cache
