+++ tests/integration_tests/test_snowflake_cache.py
@@ -80,15 +80,6 @@ def source_faker_seed_b() -> ab.Source:
     return source
 
 
-@pytest.fixture(scope="function")
-def snowflake_cache(snowflake_config) -> Generator[caches.SnowflakeCache, None, None]:
-    """Fixture to return a fresh cache."""
-    cache: caches.SnowflakeCache = caches.SnowflakeSQLCache(snowflake_config)
-    yield cache
-    # TODO: Delete cache DB file after test is complete.
-    return
-
-
 # Uncomment this line if you want to see performance trace logs.
 # You can render perf traces using the viztracer CLI or the VS Code VizTracer Extension.
 #@viztracer.trace_and_save(output_dir=".pytest_cache/snowflake_trace/")
@@ -96,11 +87,11 @@ def snowflake_cache(snowflake_config) -> Generator[caches.SnowflakeCache, None, 
 @pytest.mark.slow
 def test_faker_read_to_snowflake(
     source_faker_seed_a: ab.Source,
-    snowflake_cache: ab.SnowflakeCache,
+    new_snowflake_cache: ab.SnowflakeCache,
 ) -> None:
     """Test that the append strategy works as expected."""
     result = source_faker_seed_a.read(
-        snowflake_cache, write_strategy="replace", force_full_refresh=True
+        new_snowflake_cache, write_strategy="replace", force_full_refresh=True
     )
     assert len(list(result.cache.streams["users"])) == FAKER_SCALE_A
 
@@ -109,12 +100,12 @@ def test_faker_read_to_snowflake(
 @pytest.mark.slow
 def test_replace_strategy(
     source_faker_seed_a: ab.Source,
-    snowflake_cache: ab.SnowflakeCache,
+    new_snowflake_cache: ab.SnowflakeCache,
 ) -> None:
     """Test that the append strategy works as expected."""
     for _ in range(2):
         result = source_faker_seed_a.read(
-            snowflake_cache, write_strategy="replace", force_full_refresh=True
+            new_snowflake_cache, write_strategy="replace", force_full_refresh=True
         )
     assert len(list(result.cache.streams["users"])) == FAKER_SCALE_A
 
@@ -124,7 +115,7 @@ def test_replace_strategy(
 def test_merge_strategy(
     source_faker_seed_a: ab.Source,
     source_faker_seed_b: ab.Source,
-    snowflake_cache: ab.DuckDBCache,
+    new_snowflake_cache: ab.caches.SnowflakeCache,
 ) -> None:
     """Test that the merge strategy works as expected.
 
@@ -132,19 +123,19 @@ def test_merge_strategy(
     merge strategy.
     """
     # First run, seed A (counts should match the scale or the product count)
-    result = source_faker_seed_a.read(snowflake_cache, write_strategy="merge")
+    result = source_faker_seed_a.read(new_snowflake_cache, write_strategy="merge")
     assert len(list(result.cache.streams["users"])) == FAKER_SCALE_A
 
     # Second run, also seed A (should have same exact data, no change in counts)
-    result = source_faker_seed_a.read(snowflake_cache, write_strategy="merge")
+    result = source_faker_seed_a.read(new_snowflake_cache, write_strategy="merge")
     assert len(list(result.cache.streams["users"])) == FAKER_SCALE_A
 
     # Third run, seed B - should increase record count to the scale of B, which is greater than A.
     # TODO: See if we can reliably predict the exact number of records, since we use fixed seeds.
-    result = source_faker_seed_b.read(snowflake_cache, write_strategy="merge")
+    result = source_faker_seed_b.read(new_snowflake_cache, write_strategy="merge")
     assert len(list(result.cache.streams["users"])) == FAKER_SCALE_B
 
     # Third run, seed A again - count should stay at scale B, since A is smaller.
     # TODO: See if we can reliably predict the exact number of records, since we use fixed seeds.
-    result = source_faker_seed_a.read(snowflake_cache, write_strategy="merge")
+    result = source_faker_seed_a.read(new_snowflake_cache, write_strategy="merge")
     assert len(list(result.cache.streams["users"])) == FAKER_SCALE_B
