+++ tests/integration_tests/test_source_faker_integration.py
@@ -6,20 +6,22 @@ Since source-faker is included in dev dependencies, we can assume `source-faker`
 and available on PATH for the poetry-managed venv.
 """
 from __future__ import annotations
+
 from collections.abc import Generator
 import os
 import sys
 import shutil
 from pathlib import Path
-import typing
 
 import pytest
 import ulid
 
 
 import airbyte as ab
-from airbyte import caches
-
+from airbyte.caches.base import CacheBase
+from airbyte.caches.duckdb import DuckDBCache
+from airbyte.caches.postgres import PostgresCache
+from airbyte.caches.factories import new_local_cache, get_default_cache
 
 # Product count is always the same, regardless of faker scale.
 NUM_PRODUCTS = 100
@@ -43,12 +45,6 @@ def add_venv_bin_to_path(monkeypatch):
     # Add the bin directory to the PATH
     new_path = f"{venv_bin_path}:{os.environ['PATH']}"
     monkeypatch.setenv('PATH', new_path)
-
-
-def test_which_source_faker() -> None:
-    """Test that source-faker is available on PATH."""
-    assert shutil.which("source-faker") is not None, \
-        f"Can't find source-faker on PATH: {os.environ['PATH']}"
 
 
 @pytest.fixture(scope="function")  # Each test gets a fresh source-faker instance.
@@ -96,27 +92,26 @@ def source_faker_seed_b() -> ab.Source:
 
 
 @pytest.fixture(scope="function")
-def duckdb_cache() -> Generator[caches.DuckDBCache, None, None]:
+def duckdb_cache() -> Generator[DuckDBCache, None, None]:
     """Fixture to return a fresh cache."""
-    cache: caches.DuckDBCache = ab.new_local_cache()
+    cache: DuckDBCache = new_local_cache()
     yield cache
     # TODO: Delete cache DB file after test is complete.
     return
 
 
 @pytest.fixture(scope="function")
-def postgres_cache(new_pg_cache_config) -> Generator[caches.PostgresCache, None, None]:
+def postgres_cache(new_pg_cache) -> Generator[PostgresCache, None, None]:
     """Fixture to return a fresh cache."""
-    cache: caches.PostgresCache = caches.PostgresCache(config=new_pg_cache_config)
-    yield cache
+    yield new_pg_cache
     # TODO: Delete cache DB file after test is complete.
     return
 
 
 @pytest.fixture
 def all_cache_types(
-    duckdb_cache: ab.DuckDBCache,
-    postgres_cache: ab.PostgresCache,
+    duckdb_cache: DuckDBCache,
+    postgres_cache: PostgresCache,
 ):
     _ = postgres_cache
     return [
@@ -124,9 +119,22 @@ def all_cache_types(
         postgres_cache,
     ]
 
+
+def test_which_source_faker() -> None:
+    """Test that source-faker is available on PATH."""
+    assert shutil.which("source-faker") is not None, \
+        f"Can't find source-faker on PATH: {os.environ['PATH']}"
+
+
+def test_duckdb_cache(duckdb_cache: DuckDBCache) -> None:
+    """Test that the duckdb cache is available."""
+    assert duckdb_cache
+    assert isinstance(duckdb_cache, DuckDBCache)
+
+
 def test_faker_pks(
     source_faker_seed_a: ab.Source,
-    duckdb_cache: ab.DuckDBCache,
+    duckdb_cache: DuckDBCache,
 ) -> None:
     """Test that the append strategy works as expected."""
 
@@ -136,14 +144,14 @@ def test_faker_pks(
     assert catalog.streams[1].primary_key
 
     read_result = source_faker_seed_a.read(duckdb_cache, write_strategy="append")
-    assert read_result.cache._get_primary_keys("products") == ["id"]
-    assert read_result.cache._get_primary_keys("purchases") == ["id"]
+    assert read_result.cache.processor._get_primary_keys("products") == ["id"]
+    assert read_result.cache.processor._get_primary_keys("purchases") == ["id"]
 
 
 @pytest.mark.slow
 def test_replace_strategy(
     source_faker_seed_a: ab.Source,
-    all_cache_types: ab.DuckDBCache,
+    all_cache_types: CacheBase,
 ) -> None:
     """Test that the append strategy works as expected."""
     for cache in all_cache_types: # Function-scoped fixtures can't be used in parametrized().
@@ -158,7 +166,7 @@ def test_replace_strategy(
 @pytest.mark.slow
 def test_append_strategy(
     source_faker_seed_a: ab.Source,
-    all_cache_types: ab.DuckDBCache,
+    all_cache_types: CacheBase,
 ) -> None:
     """Test that the append strategy works as expected."""
     for cache in all_cache_types: # Function-scoped fixtures can't be used in parametrized().
@@ -174,7 +182,7 @@ def test_merge_strategy(
     strategy: str,
     source_faker_seed_a: ab.Source,
     source_faker_seed_b: ab.Source,
-    all_cache_types: ab.DuckDBCache,
+    all_cache_types: CacheBase,
 ) -> None:
     """Test that the merge strategy works as expected.
 
@@ -208,7 +216,7 @@ def test_merge_strategy(
 def test_incremental_sync(
     source_faker_seed_a: ab.Source,
     source_faker_seed_b: ab.Source,
-    duckdb_cache: ab.DuckDBCache,
+    duckdb_cache: CacheBase,
 ) -> None:
     config_a = source_faker_seed_a.get_config()
     config_b = source_faker_seed_b.get_config()
@@ -222,7 +230,7 @@ def test_incremental_sync(
     assert len(list(result1.cache.streams["purchases"])) == FAKER_SCALE_A
     assert result1.processed_records == NUM_PRODUCTS + FAKER_SCALE_A * 2
 
-    assert not duckdb_cache.get_state() == []
+    assert not duckdb_cache.processor._catalog_manager.get_state("source-faker") == []
 
     # Second run should not return records as it picks up the state and knows it's up to date.
     result2 = source_faker_seed_b.read(duckdb_cache)
@@ -243,15 +251,16 @@ def test_incremental_state_cache_persistence(
     source_faker_seed_a.set_config(config_a)
     source_faker_seed_b.set_config(config_b)
     cache_name = str(ulid.ULID())
-    cache = ab.new_local_cache(cache_name)
+    cache = new_local_cache(cache_name)
     result = source_faker_seed_a.read(cache)
     assert result.processed_records == NUM_PRODUCTS + FAKER_SCALE_A * 2
-    second_cache = ab.new_local_cache(cache_name)
+    second_cache = new_local_cache(cache_name)
     # The state should be persisted across cache instances.
     result2 = source_faker_seed_b.read(second_cache)
     assert result2.processed_records == 0
 
-    assert not second_cache.get_state() == []
+    assert second_cache.processor._catalog_manager and \
+        second_cache.processor._catalog_manager.get_state("source-faker")
     assert len(list(result2.cache.streams["products"])) == NUM_PRODUCTS
     assert len(list(result2.cache.streams["purchases"])) == FAKER_SCALE_A
 
@@ -268,8 +277,8 @@ def test_incremental_state_prefix_isolation(
     source_faker_seed_a.set_config(config_a)
     cache_name = str(ulid.ULID())
     db_path = Path(f"./.cache/{cache_name}.duckdb")
-    cache = ab.DuckDBCache(config=ab.DuckDBCacheConfig(db_path=db_path, table_prefix="prefix_"))
-    different_prefix_cache = ab.DuckDBCache(config=ab.DuckDBCacheConfig(db_path=db_path, table_prefix="different_prefix_"))
+    cache = DuckDBCache(db_path=db_path, table_prefix="prefix_")
+    different_prefix_cache = DuckDBCache(db_path=db_path, table_prefix="different_prefix_")
 
     result = source_faker_seed_a.read(cache)
     assert result.processed_records == NUM_PRODUCTS + FAKER_SCALE_A * 2
